requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyMySQL==1.1.2
python-dotenv==1.0.0
gunicorn==21.2.0
cryptography==46.0.3